% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumVar.R
\name{cumVar}
\alias{cumVar}
\title{A function to calculate a cumulative varianz for a given Vector. It calculates the
standard deviation from beginning of the vector to current position in vector}
\usage{
cumVar(v)
}
\arguments{
\item{v}{numeric or integer vector to calculate the cumulative varianz}
}
\value{
Returns vector with length of input vector which contains cumulativ varianz to given point
}
\description{
A function to calculate a cumulative varianz for a given Vector. It calculates the
standard deviation from beginning of the vector to current position in vector
}
\examples{
 v <- c(1,4,5,7,7,1,2,NA,34,NA,NA,NA,2,4,76,8,9,2,2)
 y <- cumVar(v)
 y
 y[5]
 var(v[1:5])
 y[14]
 var(v[1:14])

}
