% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_missing.R
\name{ggplot_missing}
\alias{ggplot_missing}
\title{A function to plot missing values in a dataset using ggplot}
\usage{
ggplot_missing(x)
}
\arguments{
\item{x}{data.frame or vector which should be used for the missing value plot}
}
\value{
visualises missing values for each variable in the dataset
}
\description{
A function to plot missing values in a dataset using ggplot
}
\examples{
x <- data.frame(a = rep(1,103), b = c(rep(2,53), rep(NA,20), rep(2, 30)), c = rep(NA, 103))
ggplot_missing(x)

}
