% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramGrid.R
\name{histogramGrid}
\alias{histogramGrid}
\title{A function to plot histograms for all given variables in a dataset and arrange them in a grid}
\usage{
histogramGrid(dt, vars = NA, ncols = 2, binwidth = NULL)
}
\arguments{
\item{dt}{Data.frame containing the variables to plot}

\item{vars}{Variables to use for the histogram grid}

\item{ncols}{Number of columns in histogram grid. Default: 2}

\item{binwidth}{Custom binwidth for the plots.
Currently this parameter is used for every historgram}
}
\value{
Plots a grid of histograms
}
\description{
A function to plot histograms for all given variables in a dataset and arrange them in a grid
}
\examples{


}
