% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAcuracy.R
\name{modelAccuracy}
\alias{modelAccuracy}
\title{A function to calculate the accuracy of a given classifier model}
\usage{
modelAccuracy(fit, testDT, outcomeVar)
}
\arguments{
\item{fit}{Classifier model object which should be testet. 
Currently working: rpart and random forest}

\item{testDT}{Data.frame which contains the predicting variables}

\item{outcomeVar}{Outcome variable -class- of test data set. This variable is used to test the prediction}
}
\value{
Returns a list containing the predicted outcome, the cross validation table and the accuracy
}
\description{
A function to calculate the accuracy of a given classifier model
}
\examples{


}
