% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.Sd.R
\name{pop.SD}
\alias{pop.SD}
\title{A function to calculate the population standard deviation of a vector}
\usage{
pop.SD(x)
}
\arguments{
\item{x}{vector for which to calculate the population standard deviation}
}
\value{
population standard deviation
}
\description{
A function to calculate the population standard deviation of a vector
}
\examples{
set.seed(1)
pop.SD(rnorm(100))
#compare to sample standard deviation
set.seed(1)
sd(rnorm(100))
}
