% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.Var.R
\name{pop.Var}
\alias{pop.Var}
\title{A function to calculate the population variance of a vector}
\usage{
pop.Var(x)
}
\arguments{
\item{x}{vector for which to calculate the population variance}
}
\value{
population variance
}
\description{
A function to calculate the population variance of a vector
}
\examples{
set.seed(1)
pop.Var(rnorm(100))
#compared to sample variance:
set.seed(1)
var(rnorm(100))
}
